Fix grip process finder
`ps f` was not returning the running grip process, avoiding it to be killed.

Therefore when running the grip command again it was failing due to the port being already busy (with previous process).

Fixing it also allow to remove the explicit `firefox` execution because the `-b` flag on grip command will take care of that.